Fix CircleCI issue.
It seems that when a CircleCI build is triggered by issuing a new git
tag, the CIRCLE_BRANCH env var is blank, even though there is of course
a branch on which the tag was issued. This sets the default branch name
to master if no env var is available.